+++ paddleseg/models/__init__.py
@@ -63,4 +63,4 @@ from .mobileseg import MobileSeg
 from .upernet import UPerNet
 from .sinet import SINet
 from .lraspp import LRASPP
-from .psanet import PSANet
+from .mscale_ocrnet import mscale_OCRNet
