+++ paddleseg/models/backbones/__init__.py
@@ -24,3 +24,4 @@ from .stdcnet import *
 from .lite_hrnet import *
 from .shufflenetv2 import *
 from .ghostnet import *
+from .hrnetv2_psa import *
