+++ test_tipc/configs/psanet/train_infer_python.txt
@@ -6,7 +6,7 @@ Global.use_gpu:null|null
 --precision:null
 --iters:lite_train_lite_infer=20|lite_train_whole_infer=20|whole_train_whole_infer=1000
 --save_dir:
---batch_size:lite_train_lite_infer=2|lite_train_whole_infer=2|whole_train_whole_infer=3
+--batch_size:lite_train_lite_infer=1|lite_train_whole_infer=1|whole_train_whole_infer=1
 --model_path:null
 train_model_name:best_model/model.pdparams
 train_infer_img_dir:test_tipc/data/cityscapes/cityscapes_val_5.list
@@ -27,7 +27,7 @@ null:null
 ===========================export_params===========================
 --save_dir:
 --model_path:
-norm_export:export.py --config test_tipc/configs/ccnet/ccnet_resnet101_os8_cityscapes_769x769_60k.yml
+norm_export:export.py --config test_tipc/configs/psanet/psa_hrnetv2_psa_cityscapes_1024x2048_150k.yml --input_shape 1 3 1024 2048
 quant_export:null
 fpgm_export:null
 distill_export:null
@@ -35,7 +35,7 @@ export1:null
 export2:null
 ===========================infer_params===========================
 infer_model:./test_tipc/output/psanet/model.pdparams
-infer_export:export.py --config test_tipc/configs/psanet/psa_hrnetv2_psa_cityscapes_1024x2048_150k.yml
+infer_export:export.py --config test_tipc/configs/psanet/psa_hrnetv2_psa_cityscapes_1024x2048_150k.yml --input_shape 1 3 1024 2048
 infer_quant:False
 inference:deploy/python/infer.py
 --device:cpu|gpu
@@ -49,4 +49,4 @@ inference:deploy/python/infer.py
 --save_log_path:null
 --benchmark:True
 --save_dir:
---model_name:ccnet
+--model_name:psanet
